reset version to 1.0.0, rename package to `io.realm` refactor dependencies to use the `provided` scope instead of pulling the entire plugin
