+++ app/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="realm.io">
+    package="io.realm.secureuserstore">
 </manifest>
