+++ build.gradle
@@ -2,7 +2,7 @@
 apply plugin: 'ch.netzwerg.release'
 project.ext.sdkVersion = 24
 project.ext.buildTools = '24.0.3'
-project.ext.realmVersion = '2.1.0'
+project.ext.realmVersion = '2.2.0-SNAPSHOT'
 
 buildscript {
     repositories {
@@ -11,7 +11,6 @@ buildscript {
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:2.2.0'
-        classpath 'io.realm:realm-gradle-plugin:2.1.0'
         classpath 'ch.netzwerg:gradle-release-plugin:1.2.0'                 // Manage Release tags / versioning
         classpath 'com.github.dcendents:android-maven-gradle-plugin:1.5'    // Allows you to install AAR's locally
         classpath 'org.jfrog.buildinfo:build-info-extractor-gradle:4.0.1'   // OJO integration
