+++ version.txt
@@ -1,1 +1,1 @@
-1.2.1-SNAPSHOT
+1.0.0
